Remove ScrollableGrid
All the clients have migrated to GridView.  Also, remove RenderGrid,
which isn't needed by GridView.
