+++ packages/flutter/lib/src/widgets/basic.dart
@@ -19,19 +19,13 @@ export 'package:flutter/rendering.dart' show
   CrossAxisAlignment,
   CustomClipper,
   CustomPainter,
-  FixedColumnCountGridDelegate,
   FlexFit,
   FlowDelegate,
   FlowPaintingContext,
   FractionalOffsetTween,
-  GridChildPlacement,
-  GridDelegate,
-  GridDelegateWithInOrderChildPlacement,
-  GridSpecification,
   HitTestBehavior,
   MainAxisAlignment,
   MainAxisSize,
-  MaxTileWidthGridDelegate,
   MultiChildLayoutDelegate,
   Overflow,
   PaintingContext,
